+++ Realm/RLMUser.mm
@@ -383,5 +383,8 @@ using UserProfileMember = util::Optional<std::string> (SyncUserProfile::*)() con
 - (NSString *)maxAge {
     return userProfileMemberToNSString(_userProfile.max_age());
 }
+- (NSDictionary *)metadata {
+    return (NSDictionary *)RLMConvertBsonToRLMBSON(_userProfile.data());
+}
 
 @end