+++ RealmSwift/App.swift
@@ -48,6 +48,22 @@ public typealias EmailPasswordAuth = RLMEmailPasswordAuth
  */
 public typealias UserProfile = RLMUserProfile
 
+extension UserProfile {
+    /**
+     The metadata of the user.
+     The auth provider of the user is responsible for populating this `Document`.
+    */
+    public var metadata: Document {
+        guard let rlmMetadata = self.__metadata as RLMBSON?,
+            let anyBSON = ObjectiveCSupport.convert(object: rlmMetadata),
+            case let .document(metadata) = anyBSON else {
+            return [:]
+        }
+
+        return metadata
+    }
+}
+
 /// A block type used to report an error
 public typealias EmailPasswordAuthOptionalErrorBlock = RLMEmailPasswordAuthOptionalErrorBlock
 extension EmailPasswordAuth {